requirements.txt
cvxpy
cchardet
pillow
colorama
open-clip-torch
transformers
shapely
sentence-transformers
prior
procthor
objaverse
rtree
scipy
moviepy
numpy
openai
langchain
torch
torchvision
gurobipy==10.0.3
Werkzeug==2.0.1
Flask==2.0.1
compress-pickle
compress-json
black
objathor>=0.0.7
editdistance
